+++ commands/job/fill.go
@@ -10,37 +10,6 @@ import (
 	"github.com/filecoin-project/lily/lens/lily"
 )
 
-var GapFindCmd = &cli.Command{
-	Name:  "find",
-	Usage: "find gaps in the database",
-	Flags: []cli.Flag{
-		RangeFromFlag,
-		RangeToFlag,
-	},
-	Before: func(cctx *cli.Context) error {
-		return rangeFlags.validate()
-	},
-	Action: func(cctx *cli.Context) error {
-		ctx := lotuscli.ReqContext(cctx)
-
-		api, closer, err := commands.GetAPI(ctx)
-		if err != nil {
-			return err
-		}
-		defer closer()
-
-		res, err := api.LilyGapFind(ctx, &lily.LilyGapFindConfig{
-			JobConfig: RunFlags.ParseJobConfig(),
-			To:        rangeFlags.to,
-			From:      rangeFlags.from,
-		})
-		if err != nil {
-			return err
-		}
-		return commands.PrintNewJob(os.Stdout, res)
-	},
-}
-
 var GapFillCmd = &cli.Command{
 	Name:  "fill",
 	Usage: "Fill gaps in the database",
